requirements.txt
beautifulsoup4==4.9.3
chardet==4.0.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
future==0.18.3
googlesearch-python==1.1.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
numpy==1.24.2
python-whois==0.8.0
requests==2.25.1
scikit-learn==1.2.1
scipy==1.10.1
soupsieve==2.4
threadpoolctl==3.1.0
tld==0.13
urllib3==1.26.14
Werkzeug==2.2.3
wincertstore==0.2
